pyproject.toml
[project]
name = "to-do-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-filter>=25.2",
    "django-rest-framework>=0.1.0",
    "django-stubs>=5.2.7",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.5",
    "dotenv>=0.9.9",
    "psycopg2-binary>=2.9.11",
]

[tool.basedpyright]
typeCheckingMode = "standard"
reportIncompatibleVariableOverride = false
